General steps to create a regression model
To create a regression model to predict the compressive strength of cement, We need a dataset that includes various factors that might affect the compressive strength of cement. These factors may include the cement ingredients, curing time, and curing temperature, among others.

Assuming we have a suitable dataset, we can follow the above general steps to create a regression model.